+++ src/main/java/com/ciudaddeportiva/api/model/Usuario.java
@@ -12,7 +12,7 @@ Usuario {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
+    private Long id;
 
     @Column(nullable = false, unique = true)
     private String email;
@@ -26,11 +26,13 @@ Usuario {
 
 
     // Getters y Setters
-    public int getId() {
+
+
+    public Long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
